# projects/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CropCatalogCreate": {
        "description": "Create a new crop catalog entry.",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CropCatalogCreate",
        "type": "object"
      },
      "CropCatalogRead": {
        "description": "Read a crop catalog record.",
        "properties": {
          "common_names": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Names"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CropCatalogRead",
        "type": "object"
      },
      "CropCreate": {
        "description": "Create a crop.",
        "properties": {
          "area": {
            "anyOf": [
              {
                "exclusiveMaximum": 100000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "area_unit": {
            "default": "sq_m",
            "maxLength": 50,
            "title": "Area Unit",
            "type": "string"
          },
          "crop_catalog_id": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Crop Catalog Id",
            "type": "integer"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "land_id": {
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Land Id",
            "type": "integer"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "sowing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sowing Date"
          },
          "status": {
            "default": "active",
            "maxLength": 50,
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "land_id",
          "crop_catalog_id"
        ],
        "title": "CropCreate",
        "type": "object"
      },
      "CropRead": {
        "description": "Read a crop record.",
        "properties": {
          "area": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "area_unit": {
            "default": "sq_m",
            "maxLength": 50,
            "title": "Area Unit",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crop_catalog_id": {
            "title": "Crop Catalog Id",
            "type": "integer"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "farmer_id": {
            "title": "Farmer Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "land_id": {
            "title": "Land Id",
            "type": "integer"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "sowing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sowing Date"
          },
          "status": {
            "default": "active",
            "maxLength": 50,
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "land_id",
          "crop_catalog_id",
          "id",
          "farmer_id",
          "created_at",
          "updated_at"
        ],
        "title": "CropRead",
        "type": "object"
      },
      "CropUpdate": {
        "description": "Update crop fields.",
        "properties": {
          "area": {
            "anyOf": [
              {
                "exclusiveMaximum": 100000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "area_unit": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Unit"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "sowing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sowing Date"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "CropUpdate",
        "type": "object"
      },
...
      "FarmerRead": {
        "description": "Read a farmer record (current authenticated farmer).",
        "properties": {
          "auth_uid": {
            "title": "Auth Uid",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "preferred_language": {
            "default": "en",
            "maxLength": 10,
            "title": "Preferred Language",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_role": {
            "title": "User Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "auth_uid",
          "user_role",
          "created_at",
          "updated_at"
        ],
        "title": "FarmerRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "Create a PIN account.",
        "properties": {
          "full_name": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Full Name",
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Phone",
            "type": "string"
          },
          "pin": {
            "title": "Pin",
            "type": "string"
          },
          "preferred_language": {
            "default": "en",
            "maxLength": 10,
            "title": "Preferred Language",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "pin",
          "full_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "SessionRequest": {
        "description": "Log in an existing PIN account.",
        "properties": {
          "phone": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Phone",
            "type": "string"
          },
          "pin": {
            "title": "Pin",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "pin"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "description": "A freshly issued session JWT plus the farmer it belongs to.",
        "properties": {
          "farmer": {
            "$ref": "#/components/schemas/FarmerRead"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "farmer"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MyFarm API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Create a PIN account and hand back a session JWT.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "post": {
        "description": "Verify phone + PIN, issue a session JWT.\n\nThe status code tells the login screen what to do next, so it never has\nto pre-check whether an account exists (issue #50):\n\n- **404** — no account for this phone → the client offers to register.\n- **401** — account exists, wrong PIN → the client says \"incorrect PIN\".\n- **429** — too many wrong PINs for this phone; see `Retry-After` (#308).\n- **200** — `{ token, farmer }`.",
        "operationId": "create_session_api_v1_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/whoami": {
      "get": {
        "description": "Proves the auth dependency end to end: reject anything that isn't a\nlive Firebase ID token or a valid backend session JWT; otherwise return\nexactly what was verified.",
        "operationId": "whoami_api_v1_auth_whoami_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "title": "Response Whoami Api V1 Auth Whoami Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whoami",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/crops": {
      "get": {
        "description": "List all crops for the current farmer.",
        "operationId": "list_crops_api_v1_crops_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Crops Api V1 Crops Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Crops",
        "tags": [
          "crops"
        ]
      },
      "post": {
        "description": "Create a new crop.",
        "operationId": "create_crop_api_v1_crops_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Crop",
        "tags": [
          "crops"
        ]
      }
    },
    "/api/v1/crops/{crop_id}": {
      "delete": {
        "description": "Soft-delete a crop.",
        "operationId": "delete_crop_api_v1_crops__crop_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "crop_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Crop Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Crop",
        "tags": [
          "crops"
        ]
      },
      "get": {
        "description": "Get a single crop by ID.",
        "operationId": "get_crop_api_v1_crops__crop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "crop_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Crop Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Crop",
        "tags": [
          "crops"
        ]
      },
      "patch": {
        "description": "Update a crop.",
        "operationId": "update_crop_api_v1_crops__crop_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "crop_id",
            "required": true,
            "schema": {
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Crop Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Crop",
        "tags": [
          "crops"
        ]
      }
    },
...
    "/api/v1/reference/crops": {
      "get": {
        "description": "List all available crops from the catalog.",
        "operationId": "list_crop_catalog_api_v1_reference_crops_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Crop Catalog Api V1 Reference Crops Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Crop Catalog",
        "tags": [
          "reference"
        ]
      },
      "post": {
        "description": "Create a new crop or return the existing one (case-insensitive).",
        "operationId": "create_or_get_crop_api_v1_reference_crops_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropCatalogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropCatalogRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Or Get Crop",
        "tags": [
          "reference"
        ]
      }
    },
...
  }
}